+++ @xen-orchestra/web-core/lib/packages/overlay/is-thenable.ts
@@ -1,0 +1,3 @@
+export function isThenable(value: unknown): value is PromiseLike<unknown> {
+  return typeof (value as PromiseLike<unknown> | undefined)?.then === 'function'
+}