+++ @xen-orchestra/web/src/route-map.d.ts
@@ -129,6 +129,41 @@ declare module 'vue-router/auto-routes' {
       { id: ParamValue<false> },
       | never
     >,
+    '/dev/': RouteRecordInfo<
+      '/dev/',
+      '/dev',
+      Record<never, never>,
+      Record<never, never>,
+      | never
+    >,
+    '/dev/colors': RouteRecordInfo<
+      '/dev/colors',
+      '/dev/colors',
+      Record<never, never>,
+      Record<never, never>,
+      | never
+    >,
+    '/dev/icons/': RouteRecordInfo<
+      '/dev/icons/',
+      '/dev/icons',
+      Record<never, never>,
+      Record<never, never>,
+      | never
+    >,
+    '/dev/icons/[name]': RouteRecordInfo<
+      '/dev/icons/[name]',
+      '/dev/icons/:name',
+      { name: ParamValue<true> },
+      { name: ParamValue<false> },
+      | never
+    >,
+    '/dev/token': RouteRecordInfo<
+      '/dev/token',
+      '/dev/token',
+      Record<never, never>,
+      Record<never, never>,
+      | never
+    >,
     '/host/[id]': RouteRecordInfo<
       '/host/[id]',
       '/host/:id',
@@ -353,6 +388,13 @@ declare module 'vue-router/auto-routes' {
       { id: ParamValue<false> },
       | never
     >,
+    '/vif/new': RouteRecordInfo<
+      '/vif/new',
+      '/vif/new',
+      Record<never, never>,
+      Record<never, never>,
+      | never
+    >,
     '/vm/[id]': RouteRecordInfo<
       '/vm/[id]',
       '/vm/:id',
@@ -421,6 +463,13 @@ declare module 'vue-router/auto-routes' {
       '/vm/:id/vdis',
       { id: ParamValue<true> },
       { id: ParamValue<false> },
+      | never
+    >,
+    '/vm/duplicate': RouteRecordInfo<
+      '/vm/duplicate',
+      '/vm/duplicate',
+      Record<never, never>,
+      Record<never, never>,
       | never
     >,
     '/vm/new': RouteRecordInfo<
@@ -528,6 +577,36 @@ declare module 'vue-router/auto-routes' {
     'src/pages/backup/[id]/targets.vue': {
       routes:
         | '/backup/[id]/targets'
+      views:
+        | never
+    }
+    'src/pages/dev/index.vue': {
+      routes:
+        | '/dev/'
+      views:
+        | never
+    }
+    'src/pages/dev/colors.vue': {
+      routes:
+        | '/dev/colors'
+      views:
+        | never
+    }
+    'src/pages/dev/icons/index.vue': {
+      routes:
+        | '/dev/icons/'
+      views:
+        | never
+    }
+    'src/pages/dev/icons/[name].vue': {
+      routes:
+        | '/dev/icons/[name]'
+      views:
+        | never
+    }
+    'src/pages/dev/token.vue': {
+      routes:
+        | '/dev/token'
       views:
         | never
     }
@@ -729,6 +808,12 @@ declare module 'vue-router/auto-routes' {
       views:
         | never
     }
+    'src/pages/vif/new.vue': {
+      routes:
+        | '/vif/new'
+      views:
+        | never
+    }
     'src/pages/vm/[id].vue': {
       routes:
         | '/vm/[id]'
@@ -788,6 +873,12 @@ declare module 'vue-router/auto-routes' {
     'src/pages/vm/[id]/vdis.vue': {
       routes:
         | '/vm/[id]/vdis'
+      views:
+        | never
+    }
+    'src/pages/vm/duplicate.vue': {
+      routes:
+        | '/vm/duplicate'
       views:
         | never
     }